Refactor col building and rendering
